18298 FIX HPE StoreOnce: Fix saving of checkbox in ruleset
Updating from Checkmk 2.3.0 to an patch version prior to 2.4.0p9 removed the check mark from the checkbox "Ignore TLS certificate".

If you already updated to 2.4.0 and want to ignore the TLS certificate of your HPE StoreOnce, then you have to manually check the rule again.

Prior this werk the settings of the special agent "HPE StoreOnce" has not been saved correctly and the checkbox was always unset.

If you update from 2.3.0 directly to at least 2.4.0p9 you are not affected by this Werk.

Change-Id: If98e9811eb7fad2f318985de2b3652d344fce5d2