18257 FIX rabbitmq_queues: Differentiate between queues on different vhosts
Queues that are part of different vhosts can have the same name.
Since the service item in Checkmk consisted of only the queue name, queues with the same name could not be differentiated and only the first queue seen would be discovered.

With this werk, the vhost name is added to the queue name in the form
```
<vhost>/<queue>
```
if the vhost is not the default `"/"`.

This means services monitoring queues on non-default vhosts will no longer receive data and need to be rediscovered.

SUP-24796

Change-Id: Ie388b33717d1d2c90c6d75b10aca2a5316109fcc