18065 Enable access to site apache server-status for browsers
Previously each site apache exposed it's server status through the URL
`/server-status`, which is fine for monitoring purposes. For manual inspection
however, this did not work. We now additionally expose the server status via
`/###SITE###/server-status` for this purpose.

Background: In our system apache reverse proxy setup, only the requests for
`/###SITE###/` are forward to the site apache processes. For this reason the
`/server-status` of the site apache was not accessible.

Change-Id: I309282fc45775f86333247c7267c8707bf1cf420