18426 Fix count of failed services in systemd service summary
SUP-24508

In the 'systemd service summary', failed systemd services were displayed
as such, even if they had been disabled on the system or set to 'ignore'
by the user.
This led to the service summary being misinterpreted according to the
set state, due to failed, but disabled or ignored, systemd services were
not triggering a CRIT state in the systemd service summary.

Now, disabled or ignored 'failed' systemd services are not counted as
failed services.

Change-Id: Ib3f36182a321934b811798a3f2ec8d22d2bacd7e