18258 FIX OTel: Fix incompatible ruleset stopping update
This change is relevant for users of the OpenTelemetry integration with a service monitoring rule configured on Checkmk version <= 2.4.0p5.
When updating to a higher version, this ruleset led to the update being aborted with a `KeyError`.
A possible fix was to remove the ruleset and recreated after the update.

With this werk the update is again possible without interruptions.

CMK-25001

Change-Id: Idda115128f869346edcc4736d02772ad6e87d5a3