17103 FIX Could not log in with SAML2 if signing certificate was in trust store
Signing the SAML2 authentication request was failing if the certificate
used for signing was also in the Checkmk CA trust store. The reason lies
in the way that pysaml2 invokes xmlsec1, which causes OpenSSL to try to
sign the request using the public key from the cert store.

As a workaround, we unset the SSL_CERT_FILE environment variable, so
OpenSSL doesn't find the certificate.

See also:
 - https://github.com/lsh123/xmlsec/issues/939
 - https://github.com/IdentityPython/pysaml2/issues/998

SUP-24452

Change-Id: Icf8fccefc05680b246d932a565161cf22a45786d