search: add support for alternate sorting
By default the unified sort endpoint will continue to return the results
without additional sorting. However, for usability tests for the search
results sort order, we want to expose a query param that will allow the
caller to specify an alternative sorting algorithm.

To keep it simple, we are exposing only two options for now:

  - `alphabetic`: A-Z on the title attribute.
  - `weighted_index`: custom algorithm to give more weight to different
    types of matches - this algorithm is a work-in-progress.

This code was intentionally made to be easily reverted as we may not
want to expose sorting options in the final implementation.

CMK-24563

Co-authored-by: Benjamin Knapp <benjamin.knapp@checkmk.com>
Change-Id: I25d1cb5120b58ea6eb69caf1ef79687ba4e91379