18130 Fix deletion of hosts with hostname "."
When upgrading to 2.3.0, the update process will warn you about invalid
hostnames present in your configuration. These hostnames must be removed
before you can upgrade to 2.3.0.

In the case of hostnames "." and "..", the deletion process would crash.

With this werk, we fix the deletion of these hostnames to allow the
update process to complete successfully.

SUP-23656

Change-Id: I30fb8ed221675eb79abfa80147fb2529f517a9ec