agent_prism: Fix regression in HA monitoring of KVM based AHV hypervisor setups
With Werk #18094, the Nutanix Prism special agent was updated to skip HA
(High Availability) monitoring on clusters that are not exclusively
using the Nutanix Acropolis Hypervisor (AHV). Since AHV hypervisors are
based on KVM and identify as such, the special agent effectively skipped
HA monitoring on those setups. This fix adds HA monitoring to
hypervisors identifying as KVM-based to restore the original state.

This fix includes:

- Rename: _is_only_ovh to _is_ahv_exclusive to reflect correct Nutanix
  product naming
- Add "kKvm" in addition to "kAcropolis" in the _is_only_ahv validation
- Additionally catch 500 errors on "GET /ha" requests to make sure they
  won't fail the checks execution. This is an additional safety guard to
  avoid re-introducing the original problem mentioned in werk 18094

CMK-25053

Change-Id: Ie4d55bd055c9dd2fd447eb4d703d202195fdd8a2