docker tests: Use random port on container intialization
Using a hardcoded port 9000 when intializing the container may lead
to conflicts, especially in the CI environment. Therefore we do not
set an explicit host port anymore, but use a randomized one, that
is determined by docker-py. After the container startup, this port
will be available via the ports attribute of the CheckmkApp object.

CMK-25007

Change-Id: I16b11dfa27615d38413a3f68ad99494ff054e61f