unhack plugin store
Plugins are no longer stored in a global variable,
but rather passed down the call stack.
No need anymore to maintain this global variable
(be it on class level).
We make the plugin_store a regular instance
property of the SNMPFetcher.
We now have to deserialize the plugin store before
we attempt to deserialize the fetchers.

Change-Id: I4cc5bc75cfd9ace7c8e524211da9966e0eb6f1ec