18276 FIX Linux remote alert handlers
When copying already created rule of the _Remote alert handlers (Linux)_, the copy ends up with the same SSH key as the original.

While this situation is valid and no problem in itself, it finally can lead to the problem of duplicated SSH keys in the `~/.ssh/authorized_keys` file of the target user(s) on the target system.<br>
Since this is not allowed, all entries after the first one do get ignored.<br>
This lead to nonfunctional alert handlers, showing _Not executing disallowed remote alert handler._ as the summary of the alert handler's execution in the events view.

We now fix this situation by merging the ruleset of one host with idential SSH key and target user, which results in only one SSH key entry per `authorized_keys` file.

Change-Id: Ic7be94f75d16ecd6b20433eae3981e8306744836