otel: Ignore resource attribute used for host computation in hash computation
We use set and use the attribute `cmk.host.rel_path_ts` for host name computation purposes.
Since we are setting a timestamp in this field, every incoming resource has a different hash,
and is therefor identified as different instance of that host.

By ignoring this attribute for hash computation, we are able to uniquely identify the resource properly.

Change-Id: I6c3c9e8532b5654dbda933d298f186657fb77a50