search: drop dependency on type alias
This type alias to string doesn't add any value to the unified search,
and creates an unnecessary dependency to the GUI type definitions.

If we introduce a query type again, then it will actually do something
like parsing the string and perform certain validation/escaping. In the
meantime, `str` is just fine.

Change-Id: I2a35977064ae439aa2a16346a695c662c647b571