performance test: Add plotting script and resources
The script "perftest_plot.py" can evaluate performance
test data read from disk or a database. Out of this
data, graphs can be generated, either for a single job
or as comparison of several jobs.

Data read from disk will be written to the performance
database, to archive the metrics and no have to store
the files on disk. The graphs generated will not be stored
in the database, but can always be generated dynamically
out of the database data.

CMK-24170

Change-Id: Ib624643e550263b3ce1a522de61e30955a8d83ba