chore: adjust relay protocol to manage requests and responses
The fetcher itself is able to do multiple operations depending on the payload, so for the moment we just need a task to indicate the Agent Receiver and the Relay that this task is to be proxied to the fetcher, and probably another task that will be consumed by the Relay to configure it.

Regarding the payload, all those parameters like "mode", "fetcher" and its internal fields or "snmp_plugin_store" are already given from the checkmk internals so we just need to pass those parameters. They are consumed after encoding a json from a string, so we just need a payload of type string.

The RELAY_CONFIG task can be also serialized into a json, so a string payload should work for any task.

Regarding the response payload, fetch-ad-hoc returns a json serialized. So I would just go with a string in that field.

Change-Id: I4945a652bdedf14bd2abed3705f4389ee4282186