Create relay config upon activation
Creating the relay configuration files during
*CMC* config generation is somewhat an act of
despiration. The relay *would* work with the
nagios core just fine.
However we need to take advantage of the
parallelizaion that is implemented in the CMC
config generation.

In principle, the relay config generation
should not be tied to either of the cores.

SAASDEV-5118

Change-Id: I0d51d38d87b1ac5b619146ad6716b5a86bae5de6