Remove code that writes v1 werks.
Only the 2.2.0 is still using v1 werks, but this branch comes with its
own copy of the werk tool, which should be used for working with werks
in 2.2.0 branch.
So for the master branch we no longer have to write v1 werks.

This simplifies one edge case for the upcoming "remove version from
werks" change.

CMK-19319

Change-Id: Ie9db37ddb0db479bd13397e34fcbceb496b3182d