test(search): add curated test cases for sorting
Tanja had a look at common search queries like "view" and
"notifications" and manually ordered the results in a logical way.

They are currently not being sorted in this way with our weighted index
algorithm. The future development of the algorithm should look to get as
close to this "optimal" ordering as possible.

Change-Id: I831b1084fe0540e40bc9f58e8d6c349da8c26875