search: improve weighted index search
The weighted-index sorting is now based on three tiers:

  1. Match rank: how closely the query relates to a result item.
  2. Topic rank: which result topics should be prioritized.
  3. Alphabetical: sort by title in ascending order.

Whenever, there is a tie-break in a given tier, the next tier will be
invoked to resolve the order.

I was unable to remove the xfails for the curated test cases because the
algorithm is currently unable to handle different forms of a word.

For example, the query "organize" should also match "organizes" and
"organizing", or "democracy" to its derivations: "democratic" and
"democratization". In traditional NLP, this was done with lemmatization
[1], and recently, vector databases [2] have proven to be an even more
robust solution.

However, to support this in Checkmk, it would require introducing new
dependencies, multi-language support, and for the latter, architectural
changes. So for now, this is out-of-scope.

CMK-25121

Change-Id: I560b3f2c70bd2f16daf5cf237bdf7f07137945bd

[1] https://en.wikipedia.org/wiki/Lemmatization
[2] https://en.wikipedia.org/wiki/Vector_database

Change-Id: I739648cd78e7c2b7d243a6f8659d2acec06e4969