test to detect python files without .py* extensions
CMK-25072

allowing extensions: .py, .pyi, .wsgi, .typed

using an alternative approach to only checking via shebang with
ast.parse + a lot of ignores.

Test is executed in an extra CV stage, independent of what the change
seemed to contain.
As we still need to clean ~150 files, the test is set to xfail, but
findings are written to txt files nonetheless.

Change-Id: If2404f79499a165d0e7fddabea6d6ce1771cdd2d