service discovery: restore old behavior of _disabled_ip_lookup
In some cases (e.g., during "Test Configuration" in the quick setup),
tags cannot be correctly resolved within ConfigCache.ip_stack_config, resulting in a default
ip_stack_config of IPStackConfig.IPv4.
In these cases, therefore, the function _disabled_ip_lookup is called for hosts that do not actually have an IP:
this happens for example in ConfigCache.get_host_attributes.
That is why for now we need to return None instead of raising an error

Change-Id: I2348f902f89212783eab332eaebed8ea1face7c2