18131 Fix Extreme VSP temp empty item crash
Previously, the extreme_vsp_switches_temperature check would fail with a
TypeError when it encountered SNMP data where the sensor description was
empty. This issue for example occured on devices of the 8600 series.

This werk resolves the problem by attempting to identify the sensor
using an alternative OID. If no identifying information can be found,
the check will not create a service for the sensor.

SUP-24888

Change-Id: I2ae46d3a948b84da474cd70df9570153a9565395