Register relay attribute for CME and CSE as well
Simplify HostAttributeRelay along the way. No need to make show* method
return values dynamic. The attribute is only registered for the editions
which support this attribute anyways.

With the attribute being available in CME, the attribute now is
registered in the unit test environment. Therefore the attribute tests
need to be updated.

SAASDEV-4978

Change-Id: I03b3a34aabf88d857d5e5bd380400ed457a8f832