Lazy load non-en translations in Vue
Sidestep i18n layer for english text.

On language switch, we show the english text until the
translation is loaded.

This requires us to use our own dummy i18n implementation as
vue3-gettext does not expose their interpolation logic.

The alternative would be to map all i18n hooks to () => '',
not showing any UI strings until i18n is loaded.

To support reactivity, we need to watch a ref of the
currentLanguage which only exists to make the _t hooks update.

CMK-24756

Change-Id: I96471b2a8ffd44a67a9972e6ab1809bf6638659a