18299 FIX postfix_mailq and nullmailer_mailq: Fix for mail queue size
The reported mail queue size was wrong.

Previously, the agent reported queue size in kilobytes, but the Checkmk plugin read it as bytes. This caused inaccuracies.

With this werk, the agent consistently reports the size in bytes, so the plugin reads it correctly. Your mail queue metrics will be accurate again.

Change-Id: Id177486db1bf417866b920050ba55a31a1a1552b