18466 Fix temp thresholds of Aruba 6100+ switches
The sensor thresholds were supplied for the temperature sensors of the
Aruba switches for the 6100 series. However, starting with the 6200
series, the same OIDs no longer provide the thresholds, but rather the
historical min and max values. As a result, most services for these
temperature sensors were automatically set to WARN or CRIT.

Explicit default values are now used for the Aruba 6100+ series.

Change-Id: Ie07e1f91bac95d83a856cf4b5683ae86a1a0f73c