17959 FIX REST-API: Move rule audit log messages
This werk changes the audit log entries that generate when calling the
"Move a rule to a specific location" endpoint.

Specifically, when the rule gets moved to a different folder the following
message is used:
`Moved rule ... of ruleset "..." from folder "x" to position #0 in folder "y"`

Otherwise, the following message is used (which matches the GUI):
`Moved rule ... from position #0 to #1 in ruleset "..." in folder "..."`

Previously, the REST-API incorrectly reported the move to always be to the
"top" of the folder. This is now fixed with this werk.

Change-Id: I0df432021fd34ea085ea7e74f4cc33c479948dfc
JIRA-Ref: SUP-23436