performance test: Temporary site creation workarounds
Due to issues with the current performance test environment, this
change ignores the status of an incompletely stopped site and
disables the is_mount() assertion for the sites tmp folder for the
performance test sites.

NOTE: This will be reverted once the root cause of those issues is
investigated and fixed.

CMK-25427

Change-Id: Ic3c35ab67d7cb3d1092efbcd593901d2cb07b26a