feat(vue) multitone svg icons
This icon-type is the goto icon from ux perspective:

One SVG icon with with the possibility to easily adjust the icon color.
In this PoC a Multiton icon was established.

Advantage of this approach:

2 SVG => 240 diffrent possible icons.

This can also boost performance in the future and will reduce the load
for BE rendering of icons.

Prerequesites:
- proper multitone ready svg:
  use one of the following classes:
   - icon-primary-color
   - icon-secondary-color
   - icon-tertiary-color

What does this PoC not solve:
A working migration path of the cmk-frontend-icons

Change-Id: I606374e393bd1a69beed6fe5017d152094b7bd41