18523 FIX agent_kube: Fix error when specifying multiple namespaces to monitor or exclude
The kubernetes special agent configuration allows the specification of which namespaces should be monitored.
This can be done by specifying patterns matching the namespaces to be included or excluded.

In case multiple of these patterns were specified, the agent would fail with the error message
```
[special_kube] agent_kube: error: unrecognized arguments: <PATTERN>
```
This has been fixed.

SUP-25139

Change-Id: I50e821388fbaa7503dbee80f32f4f898b164a88d