18577 REST-API: Include the downtime mode in responses
The REST-API now includes the downtime `mode` in the responses of the
`Show all schedules downtimes` and `Show downtime` endpoints.

The `mode` reflects whether the downtime is flexible. It can either be:

```json
{"type": "fixed"}
```

Or, for flexible downtimes with the flex duration in minutes:

```json
{
    "type": "flexible",
    "duration_minutes": 120
}
```

Flexible downtimes can already be created with the `duration` parameter
on the `Create a host/service related scheduled downtime` endpoints.
The behaviour there doesn't change.

Change-Id: Ic823d5f89609e711c51375fdda22e136e721ea92
JIRA-Ref: SUP-23845