prepare activating strictTemplates vueCompilerOptions (6/N)
this ignore looks harmless, but actaully this ignores all errors in the
next node (including child nodes (not present in this example))

i would have hoped to only ignore a single ts error, but this seems not
possible.

i fear this could actually weaken the typing, as all other type problems
are silenced by this comment.

Change-Id: Ic5f89f69ef67dc88434ac2a7b18398f36f45dcc2