prepare activating strictTemplates vueCompilerOptions (13/N)
this keydown.enter is an actual problem: it seems to be valid to listen
for events on unrelated parent nodes, but the typechecker does not like
this. i my opinion for a good reason: the element that generates those
events is quite far away:

FormLabels -> FormAutocompleter
FormAutocompleter -> CmkDropdown
CmkDropdown -> CmkSuggestions
CmkSuggestions -> input

the input element is the one generating the events.

On the other hand: passing those events up the chain also looks like
quite some work...

Change-Id: Id113cee386afbd9eec2fed4bd19d894e6751caf6