18179 FIX datadog: Allow arbitraty service level titles
When creating custom service levels and creating a datadog integration rule,
it may happen that the rule is not being able to be created or saved due
to the name of the service level.

The same also applies when migrating from 2.30 to 2.4.0.

SUP-25112

Change-Id: I4cf62ede997c8cbb1d2b5734c06e5b34175f8a6f
(cherry picked from commit a00c057d1892e5b8bb9af7956d1b50160f198283)