18207 SEC Fix security vulnerability in win_license.bat plugin
On Windows hosts to force the English output from the `win_license.bat` plugin,
special copying logic is used (this way, the default `slmgr.vbs` script cannot find the language files).
As the script is copied to a global, unprotected location, every user has access to edit this script.
This can be exploited for malicious intent.
To eliminate this vulnerability, the `slmgr.vbs` script is copied to the protected location
in `%SystemDrive%\ProgramData\checkmk\agent\tmp` and is deleted afterwards.

**Note:** Only users who use the Windows License plug-in are affected by this issue.

Change-Id: I795f8354b409e833174b3586baec941ad11234a8