18285 FIX activate_changes: load only the changes for the sites selected
When activting changes via the REST-API, you have the option to select
the specific sites where you want the changes to be activated. However,
the changes that are loaded, which were also returned in the response,
were for all sites.  This werk addresses the issue by only loading the
changes for the sites selected. If no sites are selected, all sites'
changes will be loaded.

CMK-25126

Change-Id: I0f0d47fdf87dc6190ea1e9f8ef2626237202fb79