changes popout: remove cached status for removed sites
We cache the activation status for all sites in the browser.
This determines what is shown when the user clicks on the menu
item `Changes`.  However, after a site was deleted, the cache
remained which meant removed sites were being shown in the changes
popout. This commit removes the cache for sites that have been
deleted when the user clicks on the `Changes` menu item.

CMK-25504

Change-Id: I416cd4736a906560610c98baeb89711f1afca32f