Clean up import-untyped errors
Mypy was reporting errors when opening the files in the editor for some
reason. All files in the repo must be linter clean, according to our
config files, to be able to work effectively.

This went unnoticed by the CI because we use rules_mypy, which behaves
differently than the UI. To make both accept the suppressions, we need
to add a few "unused-ignore" for now. Hoping that rules_mypy evolves in
a direction where such suppression are not needed anymore in the future.

Change-Id: I2d4c733ec86d9c9a741c085de189ef1c9c0fa4d6