performance test: Use executemany() when writing metrics
Since we are working in an autocommit session, using a repeated
insert with "cursor.execute()" for each row is counterproductive,
as it will do a COMMIT after each row inserted, which results in
a severe performance degradation.

Therefore we use "cursor.executemany()" when inserting the metrics,
which will only do a single COMMIT and results in a much shorter
execution time.

Change-Id: I7d210cbe8cebdf355c6a536b595a415a762eb184