18259 FIX Fixed Incorrect Limits Shown for AWS EC2 Instances
When monitoring AWS EC2 instances, the limits of various details or metrics may have been incorrect.

For example, the maximum number of vCPUs shown would be 1152 instead of the value configured on AWS.

With this fix, the correct limits are sown and thus any configured rules should work as expected.

CMK-18044

Change-Id: I01cd8d802580ff638a114f50c2a909467ae942a9