17886 FIX Support Diagnostics: Fix CLI options
The command line interface of the support diagnostics provide a way to collect
diagnostics data from the command line.
However, before this werk, the options

 - <code>--performance-graphs</code>
 - <code>--checkmk-overview</code>

led to a crash.
This werk fixes that issue by adding an additional, optional
<code>HOST</code> parameter to these:

 --<code>performance-graphs [HOST]</code>
 --<code>checkmk-overview [HOST]</code>

The <code>HOST</code> parameter provides the name of the Checkmk server among
the monitored hosts.

This werk also introduces the parameter <code>--checkmk-crashes</code>, that packs
the latest crash reports.

Change-Id: I8ecc8edd9fa060b3a233f38f8903093d6f4ed41a