Azure: Allow for passing in check_levels function
In most of the check functions in lib/azure.py, it is allowed that
check_levels is passed in, to specify which check_levels function is
used.

But, create_check_metrics_function_single() was missing the option.
Allow it to take check_levels like the others, so that v2 can be used.

Change-Id: I081669ce1dcf6938b351f6dcb130b94ab04e303c