Track UUIDs of hosts directly
Up to now, we had one single source of truth regarding
the (UUID) -> (host name) mapping (that direction).
This comes at the cost of having to look through all
the links when searching for the UUID of a given host.

To manage that complexity, the agent_registration
utils module was made a dependency of the fetchers.

By maintaining a lookup for the UUID directly, we
can loose that dependency in the upcoming commit.
We just pass the filename to the fetchers, and let
them read the UUID from that file directly.

We cannot simply pass the UUID in the fetcher
configuration, because we want changes in the
registration to take effect immediately, not
only after activation.

Change-Id: I8128c0a78f792628a77110fe2a9a956e32b30645