fetchers: trust in receiver
This removes the dependency to cmk.utils.certs from the fetchers.

Fetchers used to spontaneously create the cert store in case
the receiver failed to do so. This does not seem reasonable
to me.
a) It is the job of the agent receiver and controller to establish
trust. Let's assume they do their job (or fix them).
b) Even if we wanted to have a failsafe for this scenario,
we should check this only once, upon activation -- not per
host per minute.

Change-Id: I56975fc4b41d41f2145a7e478be82cfc9aa5f652