18286 FIX site_management: remove the 'secret' field from the site management endpoints
When calling the rest api endpoint family Site management, we
previously returned the site secret in the response. This
is a security risk and therefore we have removed it. Also,
it was possible to set the secret via the rest-api edit or
create endpoints. This is not possible via the UI, so we have
also removed this possibility from the rest-api.

CMK-25445

Change-Id: Id943cc28d5f941eb092c4d1750da796fa1b84498