Make the ICore reference passed to Query actually const.
This is necessary from a semantic POV: Livestatus threads, which execute
those queries, are not allowed to mutate the ICore, they should only observe
it.  All mutation has to be done in the main thread.

The problem here was that using a wait condition technically mutates the
ICore (namely a mutex and some condition variables), but semantically it
doesn't.  We hack around this with various "mutables".  While this is not
extremely pretty, it's what "mutable" is intended for in the end.

Change-Id: I36ba702ba971aa9509c84d468012b4e41403dc2e