lib/azure: Allow mapping a fn over metric values
Allow a function to be mapped over metric values before counting the
metric. The usecase here is, for example Azure Redis gives back a "cache
miss rate" and we want a "cache hit rate", which we can compute simply
by subtracting the value from 100. This provides an easy way to continue
using all the library abstractions in lib/azure.py but still be able to
manipulate the metric before it is consumed.

Change-Id: I6c29f3a69ec45edac6f26e8ffe64a4c69a73376b