18469 Activate changes button moved to new main menu item
The "Activate Changes" button has been moved from the top right of
title bar of the main content to a separate main navigation element.
It regularly polls for new pending changes and displays the number
of changes to be activated as a badge on the menu item.

Clicking on the new menu item no longer opens the "Activate changes"
page directly. Instead, a slideout opens that allows you to activate
pending changes without leaving the current page.

You can access the "Activate Changes" page from the slideout using the
"Open Full View" button.

Change-Id: I3fc30cfc6fd9737d08e810a54af08a81940a3980