dev-requirements: Add pytest-cov for coverage
This adds pytest-cov to give developers the option to generate coverage
reports, which are useful when writing tests to see which lines have
been hit and which haven't.

By default, pytest-cov does nothing. But it allows developers to call
`pytest --cov` (optionally with some other flags to specify the kind of
report) to get a detailed coverage report.

Change-Id: I9fb5663eece479bd811e4a95ecdfba7ff39cb305