Introduce UserPermission class
It is instanciated with the current configuration and registered
permissions and typically held for the lifetime of a HTTP request.  It
provides capabilities to determine a users permissions based on the
assigned roles and their permissions. The object uses memoizating of
certain calls to prevent repeated computations during request
processing.

This also eliminates the use of a few global request_memoize calls and
replaces them with instance level memoization.

So far the business logic was just moved to the class and existing
functions (roles_of_user, user_may, may_with_roles) create a new object
on each call to call the methods of the object. In the next step we will
pull up the instanciation of the objects and use a shared object during
the processing of a request to make the effective.

Change-Id: I81e5d19fa12ee23d202dd268687851e70b6ff1ef