diagnostics: suggest simpler API
All these ABCs make the typing needlessly complicated.
We have seen in many places that, if we aim for clean
typing in a component, we need to leave the serialization
to the user of the API of said component.

Observe that this removes the coupling between
cmk.utils.diagnostics and cmk.utils.structured_data
(the latter belonging to cmk.inventory).

Change-Id: Idab666c21b26e18156961670514650dfb45e8e73