nginx_status: factor out netstat stats extraction
Prerequisite work to extract out the actual parsing logic of the netstat
output into its own function. This makes it easier to test without
needing to patch the underlying subprocess call.

The example netstat output was derived by:

  - starting an nginx:latest (debian) container.
  - install net-tools package to get access to `netstat` utility.
  - run `os.popen("netstat -tlnp 2>/dev/null").readlines()` in Python
    REPL and copy out contents.

This doesn't cover all edge cases for dealing with netstat output, but
it's a real life example of what it could look like when nginx is
running on a host.

Change-Id: I7f9eda565dc287ecad2f36f0f4d1c24353381039