18567 nginx_status: prefer iproute2 over netstat
With this change, we add a network stats handler with the `ss` utility
from iproute2. Since `netstat` is deprecated, `ss` should take
precedence. However, if `ss` is not available, fall back to `netstat`.

The example iproute2 (ss) output was derived by:

  - starting an nginx:latest (debian) container.
  - install iproute2 package to get access to `ss` utility.
  - run `os.popen("ss -tlnp 2>/dev/null").readlines()` in Python
    REPL and copy out contents.

This doesn't cover all edge cases for dealing with ss output, but
it's a real life example of what it could look like when nginx is
running on a host.

CMK-25567

Change-Id: Ie096f49f8bb5ef5b07e4ab61286ac6d7b05ccd9f