18474 FIX Ignore login banner for mk_db2
You are affected by this Werk if you use the `mk_db2` agent plugin on
AIX.

`mk_db2.aix` use a login shell since [Werk #17480](https://checkmk.com/werk/17480).
This means that anything written to stdout in `/etc/profile` will be
part of the db2 sections. This of course will crash the check as the
checks do not expect random data included in their sections.

This Werk now applies some redirecting to remove the welcome message.

Change-Id: Ief54c294ae6c5d178688a5eb829b2f31de760c06