Azure lib: Threshold params might be nested
Allow for pulling out threshold params from a nested dictionary by
allowing to pass a lambda to the MetricData.

This makes sense here because usually these two parameters get grouped
together (they make no sense on their own), so we want to be able to go
down a level and pull the params out.

Also make levels be a required param of the threshold helper function;
it doesn't really make sense to use it without levels.

Change-Id: I667278ef5b0b6e59421e84dd3e3e5150f3ec7b5d