bakery: allow bakery plugins to reside in cmk.base.plugins.bakery
This enables us to move the bakery plug-ins to the
raw edition before migrating them to the new API
(where they would end up in the raw edition anyway).
Doing that, we can a) declutter the migration step from
the license changes, and b) properly showcase the migration
in the public git repo (as you can see a better diff then).

Change-Id: I6899fc2310f6c586de1a2963bf23936cda21b8da