18363 REST API no longer automatically migrates deprecated rule values
In earlier versions, the REST API also accepted obsolete parameters from rules.
These parameters were automatically migrated to the newer version.
The problem was that the client was not informed about what was actually written to the hard drive.
During a migration there was always a risk of newly introduced or deprecated fields.

This change requires the REST API client to use the latest known data format for the rule.
If the endpoint receives deprecated data, it reports an error.

Change-Id: I8f1cef2738a385dbd8177e7ac528f25810272dda