18568 jolokia: handle missing metrics
There is no guarantee that all the metrics will be found in the agent
output. So, it is important that we handle these cases. If at least one
metric was found, the service will be considered OK. However, if no
metrics were found in the data, the status will be UNKNOWN.

SUP-25190

Change-Id: I895acbe2c846eb7388f1ee0fb45abafbef1bd61a