18527 FIX check-http: Incorrect expected response time leading to long-running check
This werk is relevant for users using the active check `check-http` and a Checkmk version that includes [werk #16681](https://checkmk.com/werk/16681).

With [werk #16681](https://checkmk.com/werk/16681) the limits for the expected response time were scaled incorrectly, leading to large values.
Since the configured timeout is automatically increased to match the response time limits, this results in `check-http` running for a long time without finishing.

If you are affected by this werk, you need to manually check and correct your configuration. Typically, you will see days instead of seconds or milliseconds as a unit.

Change-Id: I9e46583f22ddb2f1e657988a512faebca21faacd