Windows time service check plugin
This introduces a check plugin for the Windows time service (w32time).
It does so by parsing the output of w32tm, as there is no public API
available with the same information. This comes with its own set of
issues.

Until the necessary Windows C++ agent changes land, this change also
includes a temporary PowerShell plugin, just to enable manual testing of
the check plugin. Note that this will go away and should not be used in
any kind of production environment and will not ship in a release.

CMK-25819 / CMK-22520

Change-Id: I67b7e7e61188961f1d47e78fcd28cefe6dd46235