use internal credentials on agent receiver relay tasks endpoints
We need to call checkmk rest api to find out about the relays.
This requires credentials, we used them from the call headers before.

Now we use internal credentials. This avoid the need in the header,
the calls are secure again in the planned future as we implement
mTLS authz for the relay and the site will have a unix socket to communicate with

Change-Id: Id4d22dc1bbc939555dbf0d9fb53c0a3da15e5939