Legacy tests: Make subcheck tests run
We had a number of legacy check tests that were unexpectedly not getting
run, due to an assumption that discover_legacy_checks() will return the
dotted key name that check_info gets assigned. However, it doesn't, it
always returns the full check name. This meant that any tests that were
trying to test subchecks weren't actually getting run.

The fix is a hack based on the fact that each subcheck name is of the
form "{outer_check_name}_{subcheck_name}". If a subcheck_name is given
as a key in the dataset, and a check plugin of the correct form doesn't
exist, it will error. (This doesn't happen, because the assumption
holds, but it helps prevents any typos in future regression tests.)

Also fix one test that was failing now that it is getting run.

Change-Id: I4eba89747d5ec90e27d4bae860e44e0963b34c2e