Welcome Slideout opens on Welcome Page too
Move welcome-snapin-slideout one level up, so that it's present for
every page inside the main iframe, including the Welcome Page.

This way, we don't need a complex solution to deactivate the
Welcome Snapin button when the Welcome Page is open.

This is a temporary solution until we get rid of the iframes.

The slideout has nothing to do with the header, but
moving the slidout outside of the iframe causes the slideut to
cover the snapins bar, and its links to open without main menu bar.

CMK-25762

Change-Id: I064335326a91cd4825d2e798d2a9e69d278844c7