18642 FIX check_cert URL Parameter Renamed
The check_cert active check used --url to represent a hostname to do a certificate check on.

This parameter has been renamed to --hostname.

In addition, the short form of the parameter is -H.

If you configure certificate checks through the UI then manual intervention is not required.

However, if you use the active check externally or from a script, then manual intervention is needed.

SUP-25490

Change-Id: I861d6262a16b01c4eabb92fcc5fb855db2259e19