testlib.agent_hosts: Efficiently discover services for piggybacked hosts
Instead of sequentially iterating over the list of hosts one by one and
then rescheduling services up to 3 times, it is much more efficient to
reschedule services once for all hosts and retry this several times until
the service discovery is completed. This is especially true for larger
amounts of hosts, since it parallelizes the rescheduling and does not
multiply the wait time.

Change-Id: I7ff48b1105c404fd644a432321eb1a5c328d108a