test_apache_restart_trigger: skip invalid implementation
There does not appear to be a way to trigger update hooks of Checkmk
without reinstalling apache2. However, such a reinstall breaks the
system Apache:
```
zypper install --force -y apache2
/mod_headers.so: undefined symbol: ap_set_content_type_ex
```
This is true for the site Apache, but also the system Apache. It it thus
an issue unrelated to Checkmk.

CMK-18266

Change-Id: Ifb34c455d9e110165af75d25a503cbcef61683b2